requirements.txt
akshare
llama_index=0.12.8
python-dotenv
json-repair
streamlit=1.41.1